Weather station
This station is made of an arduino mega, arduino uno, C3000 wifi shield
and sparkfun weather shield.
It measures temperature, humidity and pressure and reports them to
weather underground.
ConnectionTest1MEGA is the sketch for the mega.
Weather_Shield is the sketch for the Uno.
wiring.pdf contains wiring instructions.